requirements.txt
pre-commit==4.2.0
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.4